oopsi

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: aperron <marvin@42lausanne.ch>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 16:02:36 by aperron           #+#    #+#             */
-/*   Updated: 2023/10/24 15:29:12 by aperron          ###   ########.fr       */
+/*   Updated: 2023/10/25 16:25:33 by aperron          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	ft_bzero(void *s, size_t n)
 {
 	char	*str;
-	int		index;
+	size_t	index;
 
 	str = (char *)s;
 	index = 0;
